+++ service/sharing/model.go
@@ -1825,6 +1825,8 @@ type TableInternalAttributes struct {
 	// Storage locations of all table dependencies for shared views. Used on the
 	// recipient side for SEG (Secure Egress Gateway) whitelisting.
 	DependencyStorageLocations []string `json:"dependency_storage_locations,omitempty"`
+	// Whether the table has uniform enabled.
+	HasDeltaUniformIceberg bool `json:"has_delta_uniform_iceberg,omitempty"`
 	// Will be populated in the reconciliation response for VIEW and
 	// FOREIGN_TABLE, with the value of the parent UC entity's storage_location,
 	// following the same logic as getManagedEntityPath in